Imbens-Kalyanaraman 2012 Optimal Bandwidth Calculation

bw_ik12 calculates the Imbens-Kalyanaraman (2012) optimal bandwidth for local linear regression in regression discontinuity designs. It is based on a function in the "rddtools" package. This is an internal function and is typically not directly invoked by the user. It can be accessed using the triple colon, as in rddapp:::bw_ik12().

bw_ik12(X, Y, cutpoint = NULL, verbose = FALSE, kernel = "triangular")

Arguments

X

A numeric vector containing the running variable.

Y

A numeric vector containing the outcome variable.

cutpoint

A numeric value containing the cutpoint at which assignment to the treatment is determined. The default is 0.

verbose

A logical value indicating whether to print more information to the terminal. The default is FALSE.

kernel

A string indicating which kernel to use. Options are "triangular" (default and recommended), "rectangular", "epanechnikov", "quartic", "triweight", "tricube", and "cosine".

Value

ik_bw12 returns a numeric value specifying the optimal bandwidth.

References

Imbens, G., Kalyanaraman, K. (2012). Optimal bandwidth choice for the regression discontinuity estimator. The Review of Economic Studies, 79(3), 933-959. https://academic.oup.com/restud/article/79/3/933/1533189.

Stigler, M. and B. Quast, B (2016). rddtools: A toolbox for regression discontinuity in R.