Multiple Imputation of Multivariate Regression Discontinuity Estimation

mrd_impute estimates treatment effects in an MRDD with imputed missing values.

mrd_impute(
  formula,
  data,
  subset = NULL,
  cutpoint = NULL,
  bw = NULL,
  front.bw = NA,
  m = 10,
  k = 5,
  kernel = "triangular",
  se.type = "HC1",
  cluster = NULL,
  impute = NULL,
  verbose = FALSE,
  less = FALSE,
  est.cov = FALSE,
  est.itt = FALSE,
  local = 0.15,
  ngrid = 250,
  margin = 0.03,
  boot = NULL,
  method = c("center", "univ", "front"),
  t.design = NULL,
  stop.on.error = TRUE
)

Arguments

formula

The formula of the MRDD. This is supplied in the format of y ~ x1 + x2 for a simple sharp MRDD, or y ~ x1 + x2 | c1 + c2 for a sharp MRDD with two covariates. Fuzzy MRDD may be specified as y ~ x1 + x2 + z where x is the running variable, and z is the endogenous treatment variable. Covariates are then included in the same manner as in a sharp MRDD.

data

An optional data frame.

subset

An optional vector specifying a subset of observations to be used.

cutpoint

The cutpoint. If omitted, it is assumed to be 0.

bw

A numeric vector specifying the bandwidths at which to estimate the RD. If omitted or it is "IK12", the bandwidth is calculated using the Imbens-Kalyanaraman 2012 method. If it is "IK09", the bandwidth is calculated using the Imbens-Kalyanaraman 2009 method. Then it is estimated with that bandwidth, half that bandwidth, and twice that bandwidth. If only a single value is passed into the function, the RD will similarly be estimated at that bandwidth, half that bandwidth, and twice that bandwidth.

front.bw

A numeric vector specifying the bandwidths at which to estimate the RD for each of three effects models in the frontier method. If NA, front.bw will be determined by cross validation.

m

The number of uniformly-at-random samples to draw as search candidates for front.bw if not given.

k

An integer specifying the number of folds for cross validation to determine front.bw if not given.

kernel

A string specifying the kernel to be used in the local linear fitting. "triangular" kernel is the default and is the "correct" theoretical kernel to be used for edge estimation as in RDD (Lee and Lemieux, 2010). Other options are "rectangular", "epanechnikov", "quartic", "triweight", "tricube", "gaussian" and "cosine".

se.type

This specifies the robust SE calculation method to use. Options are, as in vcovHC, "HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5". This option is overridden by cluster.

cluster

An optional vector specifying clusters within which the errors are assumed to be correlated. This will result in reporting cluster robust SEs. This option overrides anything specified in se.type. It is suggested that data with a discrete running variable be clustered by each unique value of the running variable (Lee and Card, 2008).

impute

An optional vector specifying the imputed variables with missing values.

verbose

Will provide some additional information printed to the terminal.

less

Logical. If TRUE, return the estimates of linear and optimal, instead of linear, quadratic, cubic, optimal, half and double.

est.cov

Logical. If TRUE, the estimates of covariates will be included.

est.itt

Logical. If TRUE, the estimates of ITT will be returned.

local

The range of neighboring points around the cutoff on the standardized scale on each assignment variable, which is a positive number.

ngrid

The number of non-zero grid points on each assignment variable, which is also the number of zero grid points on each assignment variable. Value used in Wong, Steiner and Cook (2013) is 2500, which may cause long computational time.

margin

The range of grid points beyond the minimum and maximum of sample points on each assignment variable.

boot

The number of bootstrap samples to obtain standard deviation of estimates.

method

The method to estimate rd effect. Options are "center", "univ", "front".

t.design

The treatment option according to design. The 1st entry is for X1: "g" means treatment is assigned if X1 is greater than its cutoff, "geq" means treatment is assigned if X1 is greater than or equal to its cutoff, "l" means treatment is assigned if X1 is less than its cutoff, "leq" means treatment is assigned if X1 is less than or equal to its cutoff. The 2nd entry is for X2.

stop.on.error

Logical. If TRUE (the default), removes bootstraps which cause error in the integrate function, and resample till the specified number of bootstrap samples are acquired.

Value

mrd_impute returns an object of class "mrd", or "mrdi" for "front" method.

References

Stata: 64 mi estimate - Estimation using multiple imputations

Examples

x1 <- runif(300, -1, 1)
x2 <- runif(300, -1, 1)
cov <- rnorm(300)
y <- 3 + 2 * (x1 >= 0) + 3 * cov + 10 * (x2 >= 0) + rnorm(300)
imp <- rep(1:3, each = 100)
# all examples below have smaller numbers of m to keep run-time low
# centering
mrd_impute(y ~ x1 + x2 | cov, impute = imp, method = "center", t.design = c("geq", "geq"), m = 3)
#> $call
#> mrd_impute(formula = y ~ x1 + x2 | cov, m = 3, impute = imp, 
#>     method = "center", t.design = c("geq", "geq"))
#> 
#> $center
#> $center$tau_MRD
#> 
#> Call:
#> rd_est(formula = Y ~ X | cov, data = data, subset = NULL, cutpoint = 0, 
#>     bw = NULL, kernel = "triangular", se.type = "HC1", cluster = cluster, 
#>     verbose = FALSE, less = FALSE, est.cov = FALSE, est.itt = FALSE, 
#>     t.design = "leq")
#> 
#> Coefficients:
#>     Linear   Quadratic       Cubic         Opt    Half-Opt  Double-Opt  
#>      6.345       8.425       9.041       7.511       7.687       6.618  
#> 
#> 
#> 
#> attr(,"class")
#> [1] "mrd"
# univariate
mrd_impute(y ~ x1 + x2 | cov, impute = imp, method = "univ", t.design = c("geq", "geq"), m = 3)
#> $call
#> mrd_impute(formula = y ~ x1 + x2 | cov, m = 3, impute = imp, 
#>     method = "univ", t.design = c("geq", "geq"))
#> 
#> $univ
#> $univ$tau_R
#> 
#> Call:
#> rd_est(formula = Y ~ X1 | cov, data = data, subset = subset1, 
#>     cutpoint = 0, bw = NULL, kernel = "triangular", se.type = "HC1", 
#>     cluster = cluster, verbose = FALSE, less = FALSE, est.cov = FALSE, 
#>     est.itt = FALSE, t.design = "geq")
#> 
#> Coefficients:
#>     Linear   Quadratic       Cubic         Opt    Half-Opt  Double-Opt  
#>      2.419       1.816       1.567       2.253       1.850       2.373  
#> 
#> 
#> $univ$tau_M
#> 
#> Call:
#> rd_est(formula = Y ~ X2 | cov, data = data, subset = subset2, 
#>     cutpoint = 0, bw = NULL, kernel = "triangular", se.type = "HC1", 
#>     cluster = cluster, verbose = FALSE, less = FALSE, est.cov = FALSE, 
#>     est.itt = FALSE, t.design = "geq")
#> 
#> Coefficients:
#>     Linear   Quadratic       Cubic         Opt    Half-Opt  Double-Opt  
#>      9.396       9.792       9.665       9.480      10.172       9.442  
#> 
#> 
#> 
#> attr(,"class")
#> [1] "mrd"
# frontier
mrd_impute(y ~ x1 + x2 | cov, impute = imp, method = "front", t.design = c("geq", "geq"), m = 3)
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> $call
#> mrd_impute(formula = y ~ x1 + x2 | cov, m = 3, impute = imp, 
#>     method = "front", t.design = c("geq", "geq"))
#> 
#> $front
#> $front$tau_MRD
#> $call
#> mfrd_est(y = Y, x1 = X1, x2 = X2, c1 = 0, c2 = 0, t.design = c("geq", 
#> "geq"), local = 0.15, front.bw = NA, m = 3, k = 5, kernel = "triangular", 
#>     ngrid = 250, margin = 0.03, boot = NULL, cluster = NULL, 
#>     stop.on.error = TRUE)
#> 
#> $est
#>                ev1      ev2      ate    htev1    htev2    htate     tev1
#> Param     4.417508 9.420837 7.053727 3.573446 9.297282 6.728376 3.729320
#> bw        4.493366 9.273273 7.066543 4.100272 9.386221 7.031723 6.497746
#> Half-bw   5.067538 9.136878 7.351234 4.312837 9.365958 7.149061 6.930173
#> Double-bw 4.439404 9.363551 7.058468 3.726251 9.309660 6.831087 5.565995
#>               tev2     tate
#> Param     3.729319 3.729320
#> bw        6.497745 6.497745
#> Half-bw   6.930172 6.930173
#> Double-bw 5.565994 5.565994
#> 
#> $d
#>                 ev1      ev2      ate     htev1    htev2    htate      tev1
#> Param     0.7894460 1.682252 1.253604 0.6409947 1.660072 1.195197 0.6657581
#> bw        0.8033809 1.655318 1.256027 0.7328922 1.673649 1.247741 1.1598234
#> Half-bw   0.9071880 1.630774 1.308227 0.7708797 1.669556 1.268785 1.2365142
#> Double-bw 0.7935522 1.671787 1.254496 0.6677619 1.661381 1.212929 0.9918651
#>               tev2      tate
#> Param     0.665758 0.6657581
#> bw        1.159823 1.1598233
#> Half-bw   1.236514 1.2365142
#> Double-bw 0.991865 0.9918651
#> 
#> $se
#>           ev1 ev2 ate htev1 htev2 htate tev1 tev2 tate
#> Param      NA  NA  NA    NA    NA    NA   NA   NA   NA
#> bw         NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Half-bw    NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Double-bw  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 
#> $z
#>           ev1 ev2 ate htev1 htev2 htate tev1 tev2 tate
#> Param      NA  NA  NA    NA    NA    NA   NA   NA   NA
#> bw         NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Half-bw    NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Double-bw  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 
#> $df
#>           ev1 ev2 ate htev1 htev2 htate tev1 tev2 tate
#> Param      NA  NA  NA    NA    NA    NA   NA   NA   NA
#> bw         NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Half-bw    NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Double-bw  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 
#> $p
#>           ev1 ev2 ate htev1 htev2 htate tev1 tev2 tate
#> Param      NA  NA  NA    NA    NA    NA   NA   NA   NA
#> bw         NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Half-bw    NA  NA  NA    NA    NA    NA   NA   NA   NA
#> Double-bw  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 
#> $ci
#>       ev1 ev2 ate htev1 htev2 htate tev1 tev2 tate
#> 2.5%   NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 97.5%  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 2.5%   NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 97.5%  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 2.5%   NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 97.5%  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 2.5%   NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 97.5%  NA  NA  NA    NA    NA    NA   NA   NA   NA
#> 
#> $w
#>        w1        w2 
#> 0.2665397 0.1933802 
#> 
#> $front.bw
#> [1] 1.8655762 1.7030824 0.9777374
#> 
#> $impute
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "mfrdi"
#> 
#> 
#> attr(,"class")
#> [1] "mrdi"