Bandwidth Sensitivity Simulation for Multivariate Regression Discontinuity

mrd_sens_bw refits the supplied model with varying bandwidths. Other estimation parameters are held constant.

mrd_sens_bw(object, approach = c("center", "univ1", "univ2"), bws)

Arguments

object

An object returned by mrd_est or mrd_impute.

approach

A string of the approaches to be refitted, choosing from c("center", "univ1", "univ2").

bws

A positive numeric vector of the bandwidths for refitting an mrd object.

Value

mrd_sens_bw returns a dataframe containing the estimate est and standard error se for each supplemented bandwidth.

Examples

set.seed(12345)
x1 <- runif(1000, -1, 1)
x2 <- rnorm(1000, 10, 2)
cov <- rnorm(1000)
y <- 3 + 2 * x1 + 1 * x2 + 3 * cov + 10 * (x1 >= 0) + 5 * (x2 >= 10) + rnorm(1000)
mrd <- mrd_est(y ~ x1 + x2 | cov, cutpoint = c(0, 10), t.design = c("geq", "geq"))
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
mrd_sens_bw(mrd, approach = "univ1", bws = seq(0.1, 1, length.out = 5))
#>           est        se       bw        model
#> Usr  10.35439 1.1682335 0.100000    univ1-usr
#> Usr1 10.91087 0.5816926 0.325000    univ1-usr
#> Usr2 11.12184 0.4481407 0.550000    univ1-usr
#> Usr3 11.12524 0.3739633 0.775000    univ1-usr
#> Usr4 10.94819 0.3252715 1.000000    univ1-usr
#> Opt  11.10587 0.3625977 0.824974 univ1-origin