Cutoff Sensitivity Simulation for Multivariate Regression Discontinuity

mrd_sens_cutoff refits the supplemented model with varying cutoff(s). Other estimation parameters, such as the automatically calculated bandwidth, are held constant.

mrd_sens_cutoff(object, cutoffs)

Arguments

object

An object returned by mrd_est or mrd_impute.

cutoffs

A two-column numeric matrix of paired cutoff values to be used for refitting an mrd object.

Value

A dataframe which contains the estimate est and standard error se for each pairs of cutoffs (A1 and A2). A1 contains varying cutoffs on assignment 1, and A2 assignment 2.

Examples

x1 <- runif(1000, -1, 1)
x2 <- rnorm(1000, 10, 2)
cov <- rnorm(1000)
y <- 3 + 2 * x1 + 1 * x2 + 3 * cov + 10 * (x1 >= 0) + 5 * (x2 >= 10) + rnorm(1000)
mrd <- mrd_est(y ~ x1 + x2 | cov, cutpoint = c(0, 10), t.design = c("geq", "geq"))
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
mrd_sens_cutoff(mrd, expand.grid(A1 = seq(-.5, .5, length.out = 5), A2 = 10))
#>             est        se    A1 A2            model
#> 1    7.10313950 0.4432166 -0.50 10    center-linear
#> 2    2.32015929 0.6207549 -0.50 10 center-quadratic
#> 3   -0.60558285 0.6361564 -0.50 10     center-cubic
#> 4    0.63350723 0.4041430 -0.50 10   center-optimal
#> 5   -1.63083228 0.4432792 -0.50 10      center-half
#> 6    4.23301567 0.4277361 -0.50 10    center-double
#> 7   -0.82944860 0.3973498 -0.50 10     univ1-linear
#> 8   -4.55662012 0.5626808 -0.50 10  univ1-quadratic
#> 9   -1.74495403 0.6605390 -0.50 10      univ1-cubic
#> 10  -3.25727918 0.4077444 -0.50 10    univ1-optimal
#> 11  -0.77008960 0.3800129 -0.50 10       univ1-half
#> 12  -1.74043046 0.3556935 -0.50 10     univ1-double
#> 13   5.13120794 0.2354011 -0.50 10     univ2-linear
#> 14   5.12757484 0.3719448 -0.50 10  univ2-quadratic
#> 15   5.27010318 0.5199482 -0.50 10      univ2-cubic
#> 16   4.82373968 0.6611503 -0.50 10    univ2-optimal
#> 17   3.34090135 0.9448426 -0.50 10       univ2-half
#> 18   5.28484417 0.4289903 -0.50 10     univ2-double
#> 19   9.32135811 0.3948925 -0.25 10    center-linear
#> 20   6.31077206 0.5757671 -0.25 10 center-quadratic
#> 21   3.13014319 0.6548255 -0.25 10     center-cubic
#> 22   4.31793536 0.4454410 -0.25 10   center-optimal
#> 23   1.79707116 0.4610181 -0.25 10      center-half
#> 24   7.38167850 0.4119841 -0.25 10    center-double
#> 25   3.13385485 0.5194974 -0.25 10     univ1-linear
#> 26  -1.45491863 0.5536728 -0.25 10  univ1-quadratic
#> 27  -3.53043604 0.8347392 -0.25 10      univ1-cubic
#> 28  -0.35271389 0.4287791 -0.25 10    univ1-optimal
#> 29  -1.83126776 0.5947016 -0.25 10       univ1-half
#> 30   2.26221906 0.4930225 -0.25 10     univ1-double
#> 31   5.28801671 0.1918808 -0.25 10     univ2-linear
#> 32   5.50418768 0.2900365 -0.25 10  univ2-quadratic
#> 33   5.39234296 0.3935506 -0.25 10      univ2-cubic
#> 34   5.35582633 0.5342361 -0.25 10    univ2-optimal
#> 35   4.84880160 0.7472883 -0.25 10       univ2-half
#> 36   5.52564504 0.3429797 -0.25 10     univ2-double
#> 37  11.18204942 0.3133571  0.00 10    center-linear
#> 38   9.95988919 0.4412946  0.00 10 center-quadratic
#> 39   8.27608825 0.5831954  0.00 10     center-cubic
#> 40   9.20881481 0.3497096  0.00 10   center-optimal
#> 41   8.90200900 0.4081291  0.00 10      center-half
#> 42  10.37066586 0.3089282  0.00 10    center-double
#> 43  10.55473394 0.2939294  0.00 10     univ1-linear
#> 44  10.72562987 0.4608928  0.00 10  univ1-quadratic
#> 45  11.01426467 0.6657941  0.00 10      univ1-cubic
#> 46  10.65410456 0.3390368  0.00 10    univ1-optimal
#> 47  10.82379965 0.4862742  0.00 10       univ1-half
#> 48  10.57758997 0.2992609  0.00 10     univ1-double
#> 49   5.22701563 0.1641992  0.00 10     univ2-linear
#> 50   5.27597133 0.2401897  0.00 10  univ2-quadratic
#> 51   5.27944377 0.3252944  0.00 10      univ2-cubic
#> 52   5.23803421 0.4579880  0.00 10    univ2-optimal
#> 53   4.85987099 0.6682447  0.00 10       univ2-half
#> 54   5.34207267 0.2941549  0.00 10     univ2-double
#> 55   5.50555750 0.5635105  0.25 10    center-linear
#> 56   0.56503836 0.7382611  0.25 10 center-quadratic
#> 57  -3.45137170 0.9676119  0.25 10     center-cubic
#> 58   3.41080290 0.5961380  0.25 10   center-optimal
#> 59   1.49494523 0.6883051  0.25 10      center-half
#> 60   4.60077245 0.5673187  0.25 10    center-double
#> 61   3.79578656 0.5626362  0.25 10     univ1-linear
#> 62  -1.60500821 0.5522048  0.25 10  univ1-quadratic
#> 63  -4.49030870 0.7098056  0.25 10      univ1-cubic
#> 64  -0.45637522 0.4690142  0.25 10    univ1-optimal
#> 65  -3.23582291 0.5214806  0.25 10       univ1-half
#> 66   2.81792034 0.5518403  0.25 10     univ1-double
#> 67   5.07348065 0.5993722  0.25 10     univ2-linear
#> 68   3.89782860 0.8499577  0.25 10  univ2-quadratic
#> 69   3.26914269 1.1076659  0.25 10      univ2-cubic
#> 70   3.24658502 1.3824399  0.25 10    univ2-optimal
#> 71   3.24734271 1.9619371  0.25 10       univ2-half
#> 72   3.31809477 0.9851966  0.25 10     univ2-double
#> 73   3.12954557 0.5331719  0.50 10    center-linear
#> 74  -0.44551578 0.7209001  0.50 10 center-quadratic
#> 75  -0.32808961 0.9060968  0.50 10     center-cubic
#> 76   1.45049447 0.5545786  0.50 10   center-optimal
#> 77   0.08914278 0.6694249  0.50 10      center-half
#> 78   2.35525694 0.5331274  0.50 10    center-double
#> 79   0.57068645 0.4332957  0.50 10     univ1-linear
#> 80  -4.41150400 0.6022104  0.50 10  univ1-quadratic
#> 81  -3.03277126 0.7283753  0.50 10      univ1-cubic
#> 82  -2.82975446 0.4469387  0.50 10    univ1-optimal
#> 83   0.14112550 0.4308572  0.50 10       univ1-half
#> 84  -0.59002604 0.3872202  0.50 10     univ1-double
#> 85   5.29957872 0.6614110  0.50 10     univ2-linear
#> 86   4.62956652 0.9385372  0.50 10  univ2-quadratic
#> 87   3.10277051 1.1984684  0.50 10      univ2-cubic
#> 88   3.87533706 1.4974311  0.50 10    univ2-optimal
#> 89   4.02782002 2.0466265  0.50 10       univ2-half
#> 90   3.40058630 1.0530086  0.50 10     univ2-double
#> 91  11.18204942 0.3133571  0.00 10    center-linear
#> 92   9.95988919 0.4412946  0.00 10 center-quadratic
#> 93   8.27608825 0.5831954  0.00 10     center-cubic
#> 94   9.20881481 0.3497096  0.00 10   center-optimal
#> 95   8.90200900 0.4081291  0.00 10      center-half
#> 96  10.37066586 0.3089282  0.00 10    center-double
#> 97  10.55473394 0.2939294  0.00 10     univ1-linear
#> 98  10.72562987 0.4608928  0.00 10  univ1-quadratic
#> 99  11.01426467 0.6657941  0.00 10      univ1-cubic
#> 100 10.65410456 0.3390368  0.00 10    univ1-optimal
#> 101 10.82379965 0.4862742  0.00 10       univ1-half
#> 102 10.57758997 0.2992609  0.00 10     univ1-double
#> 103  5.22701563 0.1641992  0.00 10     univ2-linear
#> 104  5.27597133 0.2401897  0.00 10  univ2-quadratic
#> 105  5.27944377 0.3252944  0.00 10      univ2-cubic
#> 106  5.25027050 0.2046895  0.00 10    univ2-optimal
#> 107  5.32920768 0.2878986  0.00 10       univ2-half
#> 108  5.23447250 0.1707155  0.00 10     univ2-double