Bandwidth Sensitivity Simulation for Regression Discontinuity

rd_sens_bw refits the supplemented model with varying bandwidth. Other estimation parameters are held constant.

rd_sens_bw(object, bws)

Arguments

object

An object returned by rd_est or rd_impute.

bws

A positive numeric vector of bandwidth for refitting an rd object.

Value

A dataframe which contains the estimate est and standard error se for each supplemented bandwidth.

Examples

x <- runif(1000, -1, 1)
cov <- rnorm(1000)
y <- 3 + 2 * x + 3 * cov + 10 * (x >= 0) + rnorm(1000)
rd <- rd_est(y ~ x | cov, t.design = "geq")
rd_sens_bw(rd, bws = seq(.1, 1, length.out = 5))
#>            est        se        bw  model
#> Usr  10.227137 0.3948907 0.1000000    usr
#> Usr1  9.972594 0.2168651 0.3250000    usr
#> Usr2 10.102482 0.1732289 0.5500000    usr
#> Usr3 10.108778 0.1491913 0.7750000    usr
#> Usr4 10.050374 0.1328529 1.0000000    usr
#> Opt  10.083779 0.1438294 0.8465968 origin