Cutoff Sensitivity Simulation for Regression Discontinuity

rd_sens_cutoff refits the supplemented model with varying cutoff(s). Other estimation parameters, such as the automatically calculated bandwidth, are held constant.

rd_sens_cutoff(object, cutoffs)

Arguments

object

An object returned by rd_est or rd_impute.

cutoffs

A numeric vector of cutoff values to be used in the refitting of an rd object.

Value

A dataframe contains the estimate est and standard error se for each cutoff values (A1). Column A1 contains varying cutoffs on the assignment variable.

Examples

x <- runif(1000, -1, 1)
cov <- rnorm(1000)
y <- 3 + 2 * x + 3 * cov + 10 * (x >= 0) + rnorm(1000)
rd <- rd_est(y ~ x | cov, t.design = "geq")
rd_sens_cutoff(rd, seq(-.5, .5, length.out = 10))
#>            est        se          A1     model
#> 1   0.36427979 0.2467532 -0.50000000    linear
#> 2  -3.58666182 0.3216063 -0.50000000 quadratic
#> 3  -2.21277765 0.3653840 -0.50000000     cubic
#> 4  -2.29204885 0.2159974 -0.50000000   optimal
#> 5  -0.47795712 0.1947975 -0.50000000      half
#> 6  -0.35730169 0.2189731 -0.50000000    double
#> 7   1.61983493 0.2928694 -0.38888889    linear
#> 8  -3.25561486 0.3456652 -0.38888889 quadratic
#> 9  -3.51057497 0.5014304 -0.38888889     cubic
#> 10 -1.48719182 0.2278473 -0.38888889   optimal
#> 11 -2.01165031 0.2965447 -0.38888889      half
#> 12  0.86967622 0.2701769 -0.38888889    double
#> 13  3.04912009 0.3320350 -0.27777778    linear
#> 14 -2.01686631 0.3198172 -0.27777778 quadratic
#> 15 -4.35617554 0.4966866 -0.27777778     cubic
#> 16  0.30313167 0.2785380 -0.27777778   optimal
#> 17 -2.82065996 0.3089887 -0.27777778      half
#> 18  2.36602780 0.3219580 -0.27777778    double
#> 19  5.09077113 0.3715991 -0.16666667    linear
#> 20  0.75794862 0.4028437 -0.16666667 quadratic
#> 21 -2.96049398 0.4032648 -0.16666667     cubic
#> 22  3.26958219 0.3948565 -0.16666667   optimal
#> 23 -0.68264160 0.3147421 -0.16666667      half
#> 24  4.56936285 0.3825749 -0.16666667    double
#> 25  8.32831181 0.3208020 -0.05555556    linear
#> 26  6.66404958 0.5511797 -0.05555556 quadratic
#> 27  4.39087977 0.6778212 -0.05555556     cubic
#> 28  7.79360817 0.3995875 -0.05555556   optimal
#> 29  5.96569565 0.5927583 -0.05555556      half
#> 30  8.14912523 0.3479857 -0.05555556    double
#> 31  7.49162434 0.3731246  0.05555556    linear
#> 32  5.36167619 0.5720318  0.05555556 quadratic
#> 33  2.35294899 0.6363621  0.05555556     cubic
#> 34  6.78188357 0.4467024  0.05555556   optimal
#> 35  4.46678629 0.5907562  0.05555556      half
#> 36  7.23949351 0.4004268  0.05555556    double
#> 37  4.00948427 0.3654359  0.16666667    linear
#> 38 -0.12570125 0.3519628  0.16666667 quadratic
#> 39 -3.46277102 0.3943059  0.16666667     cubic
#> 40  2.31185466 0.3584543  0.16666667   optimal
#> 41 -1.55933580 0.2711133  0.16666667      half
#> 42  3.47577607 0.3668451  0.16666667    double
#> 43  2.06152731 0.3147350  0.27777778    linear
#> 44 -2.66691717 0.3133721  0.27777778 quadratic
#> 45 -4.59934895 0.5109283  0.27777778     cubic
#> 46 -0.22854977 0.2576511  0.27777778   optimal
#> 47 -3.02330969 0.3215746  0.27777778      half
#> 48  1.42816615 0.2992687  0.27777778    double
#> 49  0.55774867 0.2554197  0.38888889    linear
#> 50 -3.29181060 0.3154024  0.38888889 quadratic
#> 51 -3.08748309 0.4270135  0.38888889     cubic
#> 52 -1.64988805 0.2051995  0.38888889   optimal
#> 53 -1.57569811 0.2488372  0.38888889      half
#> 54 -0.03977295 0.2328447  0.38888889    double
#> 55 -0.46409419 0.2333594  0.50000000    linear
#> 56 -3.95681779 0.3244480  0.50000000 quadratic
#> 57 -2.05724705 0.3965784  0.50000000     cubic
#> 58 -2.61331559 0.2137185  0.50000000   optimal
#> 59 -0.31142275 0.1988698  0.50000000      half
#> 60 -1.04342096 0.2098249  0.50000000    double
#> 61 10.01702034 0.1223943  0.00000000    linear
#> 62 10.25868574 0.1792364  0.00000000 quadratic
#> 63 10.23540804 0.2414164  0.00000000     cubic
#> 64 10.08736253 0.1266781  0.00000000   optimal
#> 65 10.22423198 0.1723817  0.00000000      half
#> 66 10.04270809 0.1224647  0.00000000    double