Summarize the Multivariate Regression Discontinuity

summary.mrd is a summary method for class "mrd". It is based on summary.RD function in the "rdd" package.

Usage

# S3 method for mrd
summary(object, level = 0.95, digits = max(3, getOption("digits") - 3), ...)

Arguments

object

An object of class "mrd", usually a result of a call to mrd_est.

level

A numeric value between 0 and 1 specifying the confidence level for confidence intervals. The default is 0.95.

digits

A non-negative integer specifying the number of digits to display. The default is max(3, getOption("digits") - 3).

...

Additional arguments passed to summary.

Value

summary.mrd returns a list which has the following components depending on methods implemented in the "mrd" object:

center_coefficients

A matrix containing bandwidths, number of observations, estimates, SEs, confidence intervals, z-values and p-values for each estimated bandwidth and/or parametric model.

univR_coefficients

A matrix containing bandwidths, number of observations, estimates, SEs, confidence intervals, z-values and p-values for each estimated bandwidth and/or parametric model.

univM_coefficients

A matrix containing bandwidths, number of observations, estimates, SEs, confidence intervals, z-values and p-values for each estimated bandwidth and/or parametric model.

front_coefficients

A matrix containing estimates and confidence intervals (if applicable) for the complete model.

front_ht_coefficients

A matrix containing estimates and confidence intervals (if applicable) for the heterogeneous treatment model.

front_t_coefficients

A matrix containing estimates and confidence intervals (if applicable) for the treatment only model.

References

Drew Dimmery (2016). rdd: Regression Discontinuity Estimation. R package version 0.57. https://CRAN.R-project.org/package=rdd