Cutoff Sensitivity Simulation for Multivariate Regression Discontinuity

mrd_sens_cutoff refits the supplied model with varying cutoff(s). All other aspects of the model, such as the automatically calculated bandwidth, are held constant.

Usage

mrd_sens_cutoff(object, cutoffs)

Arguments

object

An object returned by mrd_est or mrd_impute.

cutoffs

A two-column numeric matrix of paired cutoff values to be used for refitting an mrd object. The first column corresponds to cutoffs for x1 and the second column corresponds to cutoffs for x2.

Value

mrd_sens_cutoff returns a dataframe containing the estimate est and standard error se for each pair of cutoffs (A1 and A2) and for each model. A1 contains varying cutoffs for assignment 1 and A2 contains varying cutoffs for assignment 2. The model column contains the approach (either centering, univariate 1, or univariate 2) for determining the cutoff and the parametric model (linear, quadratic, or cubic) or non-parametric bandwidth setting (Imbens-Kalyanaraman 2012 optimal, half, or double) used for estimation.

References

Imbens, G., Kalyanaraman, K. (2012). Optimal bandwidth choice for the regression discontinuity estimator. The Review of Economic Studies, 79(3), 933-959. https://academic.oup.com/restud/article/79/3/933/1533189.

Examples

set.seed(12345)
x1 <- runif(15000, -1, 1)
x2 <- rnorm(15000, 10, 2)
cov <- rnorm(15000)
y <- 3 + 2 * x1 + 1 * x2 + 3 * cov + 10 * (x1 >= 0) + 5 * (x2 >= 10) + rnorm(15000)
mrd <- mrd_est(y ~ x1 + x2 | cov, cutpoint = c(0, 10), t.design = c("geq", "geq"), front.bw = c(1,1,1))
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
mrd_sens_cutoff(mrd, expand.grid(A1 = seq(-.5, .5, length.out = 3), A2 = 10))
#>            est         se   A1 A2            model
#> 1   7.06133915 0.11153432 -0.5 10    center-linear
#> 2   3.43655827 0.17056849 -0.5 10 center-quadratic
#> 3   0.35721669 0.19290161 -0.5 10     center-cubic
#> 4  -0.04307852 0.10565794 -0.5 10   center-optimal
#> 5  -0.47187916 0.12930417 -0.5 10      center-half
#> 6   2.73611335 0.10805086 -0.5 10    center-double
#> 7   0.12867819 0.10650423 -0.5 10     univ1-linear
#> 8  -3.43224536 0.14729563 -0.5 10  univ1-quadratic
#> 9  -1.74886343 0.18204919 -0.5 10      univ1-cubic
#> 10 -2.10896258 0.10924905 -0.5 10    univ1-optimal
#> 11  0.03422457 0.11739343 -0.5 10       univ1-half
#> 12 -1.00269487 0.09643648 -0.5 10     univ1-double
#> 13  4.94013303 0.05661176 -0.5 10     univ2-linear
#> 14  4.91383540 0.07472065 -0.5 10  univ2-quadratic
#> 15  4.95990000 0.09314010 -0.5 10      univ2-cubic
#> 16  4.95413766 0.12289295 -0.5 10    univ2-optimal
#> 17  4.88154422 0.17759063 -0.5 10       univ2-half
#> 18  4.96404719 0.08999705 -0.5 10     univ2-double
#> 19 10.42757312 0.07712836  0.0 10    center-linear
#> 20  9.99484594 0.10562324  0.0 10 center-quadratic
#> 21  9.30936292 0.13102254  0.0 10     center-cubic
#> 22  8.88419474 0.09381808  0.0 10   center-optimal
#> 23  8.64535520 0.11531763  0.0 10      center-half
#> 24  9.72883616 0.08116446  0.0 10    center-double
#> 25 10.07092670 0.07173077  0.0 10     univ1-linear
#> 26 10.06835898 0.10804248  0.0 10  univ1-quadratic
#> 27 10.14980310 0.14487229  0.0 10      univ1-cubic
#> 28 10.06863111 0.08594703  0.0 10    univ1-optimal
#> 29 10.16970690 0.12296214  0.0 10       univ1-half
#> 30 10.07054637 0.07275143  0.0 10     univ1-double
#> 31  4.98077951 0.04421255  0.0 10     univ2-linear
#> 32  4.92673588 0.05898149  0.0 10  univ2-quadratic
#> 33  4.96005208 0.07458740  0.0 10      univ2-cubic
#> 34  5.02652809 0.10114891  0.0 10    univ2-optimal
#> 35  5.04066199 0.14583543  0.0 10       univ2-half
#> 36  5.00630748 0.07194026  0.0 10     univ2-double
#> 37  2.44804986 0.13336600  0.5 10    center-linear
#> 38  0.51723381 0.18195708  0.5 10 center-quadratic
#> 39  1.62578347 0.23092537  0.5 10     center-cubic
#> 40  1.33924349 0.15856514  0.5 10   center-optimal
#> 41  0.96680510 0.21331634  0.5 10      center-half
#> 42  2.08631721 0.13886571  0.5 10    center-double
#> 43 -0.04454592 0.10911215  0.5 10     univ1-linear
#> 44 -3.69238884 0.15657732  0.5 10  univ1-quadratic
#> 45 -1.92328723 0.19642545  0.5 10      univ1-cubic
#> 46 -2.17910574 0.11529161  0.5 10    univ1-optimal
#> 47  0.04741440 0.12172702  0.5 10       univ1-half
#> 48 -1.13056902 0.10034159  0.5 10     univ1-double
#> 49  4.83748002 0.17284241  0.5 10     univ2-linear
#> 50  4.77334108 0.23206417  0.5 10  univ2-quadratic
#> 51  4.91870922 0.29166335  0.5 10      univ2-cubic
#> 52  4.71291965 0.39674721  0.5 10    univ2-optimal
#> 53  4.62150220 0.56674025  0.5 10       univ2-half
#> 54  4.87688553 0.28285618  0.5 10     univ2-double
#> 55 10.42757312 0.07712836  0.0 10    center-linear
#> 56  9.99484594 0.10562324  0.0 10 center-quadratic
#> 57  9.30936292 0.13102254  0.0 10     center-cubic
#> 58  8.88419474 0.09381808  0.0 10   center-optimal
#> 59  8.64535520 0.11531763  0.0 10      center-half
#> 60  9.72883616 0.08116446  0.0 10    center-double
#> 61 10.07092670 0.07173077  0.0 10     univ1-linear
#> 62 10.06835898 0.10804248  0.0 10  univ1-quadratic
#> 63 10.14980310 0.14487229  0.0 10      univ1-cubic
#> 64 10.06863111 0.08594703  0.0 10    univ1-optimal
#> 65 10.16970690 0.12296214  0.0 10       univ1-half
#> 66 10.07054637 0.07275143  0.0 10     univ1-double
#> 67  4.98077951 0.04421255  0.0 10     univ2-linear
#> 68  4.92673588 0.05898149  0.0 10  univ2-quadratic
#> 69  4.96005208 0.07458740  0.0 10      univ2-cubic
#> 70  4.95184584 0.05197190  0.0 10    univ2-optimal
#> 71  4.99901511 0.06841120  0.0 10       univ2-half
#> 72  4.96583948 0.04559202  0.0 10     univ2-double