Cutoff Sensitivity Simulation for Multivariate Regression Discontinuity

mrd_sens_cutoff refits the supplied model with varying cutoff(s). Other estimation parameters, such as the automatically calculated bandwidth, are held constant.

mrd_sens_cutoff(object, cutoffs)

Arguments

object

An object returned by mrd_est or mrd_impute.

cutoffs

A two-column numeric matrix of paired cutoff values to be used for refitting an mrd object.

Value

mrd_sens_cutoff returns a dataframe containing the estimate est and standard error se for each pairs of cutoffs (A1 and A2). A1 contains varying cutoffs for assignment 1 and A2 contains varying cutoffs for assignment 2.

Examples

set.seed(12345)
x1 <- runif(1000, -1, 1)
x2 <- rnorm(1000, 10, 2)
cov <- rnorm(1000)
y <- 3 + 2 * x1 + 1 * x2 + 3 * cov + 10 * (x1 >= 0) + 5 * (x2 >= 10) + rnorm(1000)
mrd <- mrd_est(y ~ x1 + x2 | cov, cutpoint = c(0, 10), t.design = c("geq", "geq"))
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
#> Warning: collapsing to unique 'x' values
mrd_sens_cutoff(mrd, expand.grid(A1 = seq(-.5, .5, length.out = 5), A2 = 10))
#>            est        se    A1 A2            model
#> 1    6.8847593 0.4633947 -0.50 10    center-linear
#> 2    2.9281293 0.6146645 -0.50 10 center-quadratic
#> 3    0.4035069 0.7002419 -0.50 10     center-cubic
#> 4   -1.2246745 0.4515216 -0.50 10   center-optimal
#> 5   -0.2777913 0.5475007 -0.50 10      center-half
#> 6    1.7956124 0.4347646 -0.50 10    center-double
#> 7   -0.4062718 0.4383258 -0.50 10     univ1-linear
#> 8   -4.3336221 0.6166704 -0.50 10  univ1-quadratic
#> 9   -2.5951449 0.7964317 -0.50 10      univ1-cubic
#> 10  -3.0065180 0.4434336 -0.50 10    univ1-optimal
#> 11  -0.2924274 0.4629300 -0.50 10       univ1-half
#> 12  -1.6987637 0.3953229 -0.50 10     univ1-double
#> 13   4.9056863 0.2072438 -0.50 10     univ2-linear
#> 14   4.7193419 0.2483900 -0.50 10  univ2-quadratic
#> 15   4.7362439 0.3150560 -0.50 10      univ2-cubic
#> 16   4.8059388 0.3549255 -0.50 10    univ2-optimal
#> 17   4.6979063 0.3894803 -0.50 10       univ2-half
#> 18   4.8879584 0.2899357 -0.50 10     univ2-double
#> 19   9.1249231 0.3918913 -0.25 10    center-linear
#> 20   7.0358478 0.5255029 -0.25 10 center-quadratic
#> 21   5.3222964 0.6792858 -0.25 10     center-cubic
#> 22   2.8334763 0.5038109 -0.25 10   center-optimal
#> 23   0.2431821 0.5824295 -0.25 10      center-half
#> 24   5.8302987 0.4690176 -0.25 10    center-double
#> 25   3.6317984 0.5824166 -0.25 10     univ1-linear
#> 26  -0.9408079 0.6319657 -0.25 10  univ1-quadratic
#> 27  -3.4521382 0.8794861 -0.25 10      univ1-cubic
#> 28  -0.8535589 0.5125043 -0.25 10    univ1-optimal
#> 29  -2.1702267 0.6589783 -0.25 10       univ1-half
#> 30   2.5151616 0.5776248 -0.25 10     univ1-double
#> 31   4.9505418 0.1833179 -0.25 10     univ2-linear
#> 32   5.0091259 0.2308199 -0.25 10  univ2-quadratic
#> 33   5.1308284 0.2858850 -0.25 10      univ2-cubic
#> 34   5.3367005 0.3468153 -0.25 10    univ2-optimal
#> 35   4.8114566 0.3654159 -0.25 10       univ2-half
#> 36   5.3428165 0.2731284 -0.25 10     univ2-double
#> 37  11.0754510 0.3191325  0.00 10    center-linear
#> 38  10.6340007 0.4201082  0.00 10 center-quadratic
#> 39  10.3514707 0.5461452  0.00 10     center-cubic
#> 40   9.3993911 0.3807504  0.00 10   center-optimal
#> 41   9.4285404 0.5063189  0.00 10      center-half
#> 42  10.2831952 0.3330933  0.00 10    center-double
#> 43  10.5846002 0.2876699  0.00 10     univ1-linear
#> 44  11.4707113 0.4525852  0.00 10  univ1-quadratic
#> 45  11.0841187 0.6000191  0.00 10      univ1-cubic
#> 46  11.1058669 0.3625977  0.00 10    univ1-optimal
#> 47  10.9673369 0.5058807  0.00 10       univ1-half
#> 48  10.7254182 0.2958371  0.00 10     univ1-double
#> 49   5.0774755 0.1613573  0.00 10     univ2-linear
#> 50   5.0799220 0.2079335  0.00 10  univ2-quadratic
#> 51   5.1167284 0.2588334  0.00 10      univ2-cubic
#> 52   5.2628545 0.3218006  0.00 10    univ2-optimal
#> 53   5.0055656 0.3935608  0.00 10       univ2-half
#> 54   5.3188818 0.2485518  0.00 10     univ2-double
#> 55   4.9755413 0.5573096  0.25 10    center-linear
#> 56   1.5796894 0.7163851  0.25 10 center-quadratic
#> 57   0.3722034 0.9035555  0.25 10     center-cubic
#> 58   2.4188329 0.6418796  0.25 10   center-optimal
#> 59   0.1681872 0.7991646  0.25 10      center-half
#> 60   4.1252710 0.5759858  0.25 10    center-double
#> 61   2.7489037 0.5204759  0.25 10     univ1-linear
#> 62  -1.3869267 0.4891949  0.25 10  univ1-quadratic
#> 63  -3.3484050 0.6627777  0.25 10      univ1-cubic
#> 64  -1.2239790 0.3878575  0.25 10    univ1-optimal
#> 65  -1.7648484 0.4939657  0.25 10       univ1-half
#> 66   1.7056107 0.4840635  0.25 10     univ1-double
#> 67   4.6850572 0.6529152  0.25 10     univ2-linear
#> 68   4.7864847 0.8599781  0.25 10  univ2-quadratic
#> 69   5.7161804 1.0588102  0.25 10      univ2-cubic
#> 70   6.3271897 1.3045510  0.25 10    univ2-optimal
#> 71   5.6690921 1.6128060  0.25 10       univ2-half
#> 72   5.3180571 1.0312922  0.25 10     univ2-double
#> 73   2.4466751 0.5073562  0.50 10    center-linear
#> 74   0.6223124 0.6523347  0.50 10 center-quadratic
#> 75   2.3454819 0.8083939  0.50 10     center-cubic
#> 76   0.8951054 0.6367617  0.50 10   center-optimal
#> 77   1.2029813 0.8225216  0.50 10      center-half
#> 78   2.0247158 0.5311896  0.50 10    center-double
#> 79  -0.3060398 0.4256132  0.50 10     univ1-linear
#> 80  -3.1011407 0.6547519  0.50 10  univ1-quadratic
#> 81  -1.5665685 0.8917970  0.50 10      univ1-cubic
#> 82  -1.7361964 0.4665950  0.50 10    univ1-optimal
#> 83   0.4539768 0.5234849  0.50 10       univ1-half
#> 84  -1.2354987 0.4041570  0.50 10     univ1-double
#> 85   4.4105824 0.6937743  0.50 10     univ2-linear
#> 86   4.2427715 0.9092893  0.50 10  univ2-quadratic
#> 87   4.3609328 1.1313003  0.50 10      univ2-cubic
#> 88   5.1037715 1.4559238  0.50 10    univ2-optimal
#> 89   3.4728684 2.0077789  0.50 10       univ2-half
#> 90   4.0332410 1.1046053  0.50 10     univ2-double
#> 91  11.0754510 0.3191325  0.00 10    center-linear
#> 92  10.6340007 0.4201082  0.00 10 center-quadratic
#> 93  10.3514707 0.5461452  0.00 10     center-cubic
#> 94   9.3993911 0.3807504  0.00 10   center-optimal
#> 95   9.4285404 0.5063189  0.00 10      center-half
#> 96  10.2831952 0.3330933  0.00 10    center-double
#> 97  10.5846002 0.2876699  0.00 10     univ1-linear
#> 98  11.4707113 0.4525852  0.00 10  univ1-quadratic
#> 99  11.0841187 0.6000191  0.00 10      univ1-cubic
#> 100 11.1058669 0.3625977  0.00 10    univ1-optimal
#> 101 10.9673369 0.5058807  0.00 10       univ1-half
#> 102 10.7254182 0.2958371  0.00 10     univ1-double
#> 103  5.0774755 0.1613573  0.00 10     univ2-linear
#> 104  5.0799220 0.2079335  0.00 10  univ2-quadratic
#> 105  5.1167284 0.2588334  0.00 10      univ2-cubic
#> 106  5.1060010 0.1929948  0.00 10    univ2-optimal
#> 107  5.3100372 0.2452996  0.00 10       univ2-half
#> 108  5.0726866 0.1678915  0.00 10     univ2-double