Cutoff Sensitivity Simulation for Regression Discontinuity

rd_sens_cutoff refits the supplied model with varying cutoff(s). All other aspects of the model, such as the automatically calculated bandwidth, are held constant.

Usage

rd_sens_cutoff(object, cutoffs)

Arguments

object

An object returned by rd_est or rd_impute.

cutoffs

A numeric vector of cutoff values to be used for refitting an rd object.

Value

rd_sens_cutoff returns a dataframe containing the estimate est and standard error se

for each cutoff value (A1). Column A1 contains varying cutoffs on the assignment variable. The model column contains the parametric model (linear, quadratic, or cubic) or non-parametric bandwidth setting (Imbens-Kalyanaraman 2012 optimal, half, or double) used for estimation.

References

Imbens, G., Kalyanaraman, K. (2012). Optimal bandwidth choice for the regression discontinuity estimator. The Review of Economic Studies, 79(3), 933-959. https://academic.oup.com/restud/article/79/3/933/1533189.

Examples

set.seed(12345)
x <- runif(1000, -1, 1)
cov <- rnorm(1000)
y <- 3 + 2 * x + 3 * cov + 10 * (x >= 0) + rnorm(1000)
rd <- rd_est(y ~ x | cov, t.design = "geq")
rd_sens_cutoff(rd, seq(-.5, .5, length.out = 10))
#>            est        se          A1     model
#> 1   0.07969423 0.2444090 -0.50000000    linear
#> 2  -3.56425289 0.3484931 -0.50000000 quadratic
#> 3  -2.14631083 0.4574047 -0.50000000     cubic
#> 4  -2.25549684 0.2466853 -0.50000000   optimal
#> 5   0.08617364 0.2389141 -0.50000000      half
#> 6  -1.05563755 0.2115650 -0.50000000    double
#> 7   1.13036935 0.2641814 -0.38888889    linear
#> 8  -3.01554092 0.2816176 -0.38888889 quadratic
#> 9  -3.15717647 0.3541280 -0.38888889     cubic
#> 10 -2.29963553 0.2109532 -0.38888889   optimal
#> 11 -0.44725446 0.2049669 -0.38888889      half
#> 12 -0.01664455 0.2244850 -0.38888889    double
#> 13  3.14369677 0.3310038 -0.27777778    linear
#> 14 -1.66674376 0.3016680 -0.27777778 quadratic
#> 15 -3.91768354 0.4345337 -0.27777778     cubic
#> 16 -1.14834321 0.2357961 -0.27777778   optimal
#> 17 -2.00224706 0.3039055 -0.27777778      half
#> 18  2.03843241 0.3113976 -0.27777778    double
#> 19  5.23948286 0.3804960 -0.16666667    linear
#> 20  0.87627157 0.4205371 -0.16666667 quadratic
#> 21 -3.06449227 0.3957340 -0.16666667     cubic
#> 22  1.68951550 0.3927183 -0.16666667   optimal
#> 23 -2.46201237 0.3017299 -0.16666667      half
#> 24  4.38509932 0.3993719 -0.16666667    double
#> 25  8.60263592 0.3126928 -0.05555556    linear
#> 26  6.95763696 0.5682479 -0.05555556 quadratic
#> 27  4.64096453 0.7448366 -0.05555556     cubic
#> 28  7.43246057 0.4904542 -0.05555556   optimal
#> 29  5.05961343 0.7054854 -0.05555556      half
#> 30  8.32612885 0.3574042 -0.05555556    double
#> 31  7.54145063 0.3672640  0.05555556    linear
#> 32  5.06056419 0.5491753  0.05555556 quadratic
#> 33  2.26934794 0.5770854  0.05555556     cubic
#> 34  5.69759456 0.5048512  0.05555556   optimal
#> 35  2.76292037 0.5476351  0.05555556      half
#> 36  7.09121133 0.4075403  0.05555556    double
#> 37  4.73711356 0.3838563  0.16666667    linear
#> 38  0.15873621 0.3938900  0.16666667 quadratic
#> 39 -3.58197805 0.4361977  0.16666667     cubic
#> 40  0.94210596 0.3604900  0.16666667   optimal
#> 41 -2.74438672 0.3349725  0.16666667      half
#> 42  3.83493989 0.3930972  0.16666667    double
#> 43  2.44021253 0.3070641  0.27777778    linear
#> 44 -2.01011054 0.2888192  0.27777778 quadratic
#> 45 -3.33888035 0.4000231  0.27777778     cubic
#> 46 -1.47547195 0.2187244  0.27777778   optimal
#> 47 -1.35816810 0.2691661  0.27777778      half
#> 48  1.35284668 0.2765852  0.27777778    double
#> 49  0.74623053 0.2460348  0.38888889    linear
#> 50 -3.16706426 0.2664022  0.38888889 quadratic
#> 51 -2.74805736 0.3310677  0.38888889     cubic
#> 52 -2.38293452 0.1992077  0.38888889   optimal
#> 53 -0.27524311 0.1742558  0.38888889      half
#> 54 -0.36324795 0.2028941  0.38888889    double
#> 55 -0.32689908 0.2209892  0.50000000    linear
#> 56 -3.41397231 0.2968565  0.50000000 quadratic
#> 57 -1.39376696 0.3504643  0.50000000     cubic
#> 58 -2.05049172 0.2155820  0.50000000   optimal
#> 59 -0.02753801 0.1967451  0.50000000      half
#> 60 -1.34703640 0.1892384  0.50000000    double
#> 61 10.04882040 0.1190290  0.00000000    linear
#> 62 10.06839238 0.1780133  0.00000000 quadratic
#> 63  9.82936267 0.2253520  0.00000000     cubic
#> 64 10.04667463 0.1413530  0.00000000   optimal
#> 65  9.89572808 0.1885718  0.00000000      half
#> 66 10.05245913 0.1199932  0.00000000    double